Merge reference databases

Merge several reference database by common fields.

refdb_merge(..., keep = "fields_all")

Arguments

...

reference databases (tibbles).

keep

determines which columns to keep. Can be "fields_all" (default), "fields_shared" or "all" (see Details).

Value

a merged reference database (tibble).

Details

Columns are merged only if they are associated to the same field.

The keep argument determines which columns are returned as follow. "fields_all" (the default) returns all the fields existing in all the reference databases. "fields_shared" returns only the fields shared by all the reference databases. "all" returns all the columns of all the databases. Columns which are not associated to a field are not merged and are prefixed with the name of the object they originated from.

Examples

lib_1 <- read.csv(system.file("extdata", "baetidae_bold.csv", package = "refdb"))
lib_1 <- refdb_set_fields_BOLD(lib_1)
lib_2 <- lib_1
refdb_merge(lib_1, lib_2)
#>    source sequenceID phylum_name class_name    order_name family_name
#> 1    BOLD    7208675  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 2    BOLD    7208674  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 3    BOLD   11680869  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 4    BOLD   11680870  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 5    BOLD   11680871  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 6    BOLD   11680872  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 7    BOLD   11680873  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 8    BOLD   11680874  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 9    BOLD    9698877  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 10   BOLD    9698882  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 11   BOLD    9698874  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 12   BOLD    9698867  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 13   BOLD    9698881  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 14   BOLD    9698879  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 15   BOLD    9698876  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 16   BOLD    9698878  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 17   BOLD    9698868  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 18   BOLD    9698869  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 19   BOLD    9698870  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 20   BOLD    9698866  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 21   BOLD    9698873  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 22   BOLD    9698880  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 23   BOLD    9698871  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 24   BOLD    9698884  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 25   BOLD    9698872  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 26   BOLD    9698875  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 27   BOLD    9698883  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 28   BOLD    9698885  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 29   BOLD    9698887  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 30   BOLD    9698886  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 31   BOLD    7208675  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 32   BOLD    7208674  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 33   BOLD   11680869  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 34   BOLD   11680870  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 35   BOLD   11680871  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 36   BOLD   11680872  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 37   BOLD   11680873  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 38   BOLD   11680874  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 39   BOLD    9698877  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 40   BOLD    9698882  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 41   BOLD    9698874  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 42   BOLD    9698867  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 43   BOLD    9698881  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 44   BOLD    9698879  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 45   BOLD    9698876  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 46   BOLD    9698878  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 47   BOLD    9698868  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 48   BOLD    9698869  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 49   BOLD    9698870  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 50   BOLD    9698866  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 51   BOLD    9698873  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 52   BOLD    9698880  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 53   BOLD    9698871  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 54   BOLD    9698884  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 55   BOLD    9698872  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 56   BOLD    9698875  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 57   BOLD    9698883  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 58   BOLD    9698885  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 59   BOLD    9698887  Arthropoda    Insecta Ephemeroptera    Baetidae
#> 60   BOLD    9698886  Arthropoda    Insecta Ephemeroptera    Baetidae
#>    subfamily_name genus_name     species_name
#> 1        Baetinae     Cloeon  Cloeon dipterum
#> 2        Baetinae     Cloeon  Cloeon dipterum
#> 3        Baetinae     Baetis   Baetis alpinus
#> 4        Baetinae     Baetis   Baetis alpinus
#> 5        Baetinae     Baetis   Baetis alpinus
#> 6        Baetinae     Baetis   Baetis alpinus
#> 7        Baetinae     Baetis   Baetis alpinus
#> 8        Baetinae     Baetis   Baetis alpinus
#> 9        Baetinae     Baetis   Baetis alpinus
#> 10       Baetinae     Baetis   Baetis alpinus
#> 11       Baetinae     Baetis   Baetis alpinus
#> 12       Baetinae     Baetis   Baetis alpinus
#> 13       Baetinae     Baetis   Baetis alpinus
#> 14       Baetinae     Baetis   Baetis alpinus
#> 15       Baetinae     Baetis   Baetis alpinus
#> 16       Baetinae     Baetis   Baetis alpinus
#> 17       Baetinae     Baetis   Baetis alpinus
#> 18       Baetinae     Baetis   Baetis alpinus
#> 19       Baetinae     Baetis   Baetis alpinus
#> 20       Baetinae     Baetis   Baetis alpinus
#> 21       Baetinae     Baetis   Baetis alpinus
#> 22       Baetinae     Baetis   Baetis alpinus
#> 23       Baetinae     Baetis   Baetis alpinus
#> 24       Baetinae     Baetis   Baetis alpinus
#> 25       Baetinae     Baetis   Baetis alpinus
#> 26       Baetinae     Baetis   Baetis alpinus
#> 27       Baetinae     Baetis   Baetis alpinus
#> 28       Baetinae     Baetis Baetis melanonyx
#> 29       Baetinae     Baetis Baetis melanonyx
#> 30       Baetinae     Baetis Baetis melanonyx
#> 31       Baetinae     Cloeon  Cloeon dipterum
#> 32       Baetinae     Cloeon  Cloeon dipterum
#> 33       Baetinae     Baetis   Baetis alpinus
#> 34       Baetinae     Baetis   Baetis alpinus
#> 35       Baetinae     Baetis   Baetis alpinus
#> 36       Baetinae     Baetis   Baetis alpinus
#> 37       Baetinae     Baetis   Baetis alpinus
#> 38       Baetinae     Baetis   Baetis alpinus
#> 39       Baetinae     Baetis   Baetis alpinus
#> 40       Baetinae     Baetis   Baetis alpinus
#> 41       Baetinae     Baetis   Baetis alpinus
#> 42       Baetinae     Baetis   Baetis alpinus
#> 43       Baetinae     Baetis   Baetis alpinus
#> 44       Baetinae     Baetis   Baetis alpinus
#> 45       Baetinae     Baetis   Baetis alpinus
#> 46       Baetinae     Baetis   Baetis alpinus
#> 47       Baetinae     Baetis   Baetis alpinus
#> 48       Baetinae     Baetis   Baetis alpinus
#> 49       Baetinae     Baetis   Baetis alpinus
#> 50       Baetinae     Baetis   Baetis alpinus
#> 51       Baetinae     Baetis   Baetis alpinus
#> 52       Baetinae     Baetis   Baetis alpinus
#> 53       Baetinae     Baetis   Baetis alpinus
#> 54       Baetinae     Baetis   Baetis alpinus
#> 55       Baetinae     Baetis   Baetis alpinus
#> 56       Baetinae     Baetis   Baetis alpinus
#> 57       Baetinae     Baetis   Baetis alpinus
#> 58       Baetinae     Baetis Baetis melanonyx
#> 59       Baetinae     Baetis Baetis melanonyx
#> 60       Baetinae     Baetis Baetis melanonyx
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           nucleotides
#> 1                                                 TACTTTATACTTCATTTTTGGTGCATGAGCTGGAATAGTGGGCACTTCTTTGAGTTTATTAATTCGTGCAGAGCTAGGTAATCCTGGTTCTTTAATTGGTGATGACCAGATTTATAATGTTATTGTTACTGCCCATGCTTTTATTATAATTTTCTTTATAGTTATACCAATTATGATTGGTGGTTTTGGTAATTGATTAGTTCCTCTTATATTAGGAGCTCCTGATATAGCTTTTCCTCGTATAAATAATATAAGTTTTTGATTATTGCCTCCTTCATTGACTTTATTAGTATCAAGTAGTTTAGTAGATATAGGAGCTGGTACGGGTTGAACTGTATATCCTCCTTTAGCGGCTAATATTGCTCATGGAGGGTCATCAGTTGATTACGCTATTTTTTCTTTACATTTAGCTGGGGTATCTTCTATTTTAGGTGCTGTAAATTTTATTACAACAGTAATTAACATGCGTAGCCCTGGTATAACTTTAGATCGTATTCCTTTATTTGTATGATCTGTTGTAATTACTGCTGTGTTATTGCTTTTATCGCTACCTGTATTAGCCGGTGCTATTACTATACTTTTAACCGATCGTAATCTTAATACTTCTTTTT
#> 2     TACTTTATACTTCATTTTTGGTGCATGAGCTGGTATAGTGGGCACTTCATTGAGTTTATTAATTCGTGCGGAACTAGGGAATCCTGGATCTTTAATTGGTGATGATCAAATTTATAATGTTATTGTTACTGCTCACGCTTTTATTATAATTTTCTTTATGGTTATACCAATCATGATTGGAGGTTTTGGTAATTGATTAGTACCTCTTATATTAGGAGCACCTGATATGGCCTTCCCTCGTATAAATAATATAAGTTTTTGATTATTACCTCCTTCATTAACTTTATTAGTATCAAGTAGTTTAGTTGATATGGGAGCCGGTACAGGCTGAACAGTTTATCCTCCTTTAGCTGCAAATATTGCTCATGGAGGGTCATCCGTTGATTATGCTATTTTTTCTTTACATTTAGCTGGGGTATCTTCTATTTTAGGTGCTGTAAATTTTATTACAACAGTAATTAATATGCGTAGTCCGGGTATAACTTTAGATCGAATTCCTTTATTTGTATGATCTGTTGTAATTACTGCTGTTTTATTATTATTATCACTACCTGTGTTAGCTGGTGCTATTACTATACTTTTAACTGATCGTAATCTTAATACTTCTTTTTTTGATCCTGCTGGAGGGGGTGATCCTATTTTATACCAACATTTA
#> 3                                               TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTTTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 4                                               TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCACCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTATGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCCCTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 5                                               TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTTCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCGCTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 6                                               TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATCATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTATATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCGCTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 7                                               TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCGCTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 8                                               TGGTGGGGACATCGCTCAGTCTTTTAATTCGAGCCGAGCTTGGTAACCCTGGGTCCTTAATTGGTGACGACCAGATTTACAACGTTATTGTTACTGCTCACGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGGGGGTTTGGTAATTGGCTTGTTCCCCTTATATTAGGGGCCCCCGACATAGCATTTCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCTCCTTCCCTTACACTACTTGTGTCGAGCAGTGTGGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCGCATGGGGGGGCTTCAGTGGATTTCGCAATTTTTTCGTTACACCTGGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAACTTCATTACAACTGTAATTAATATGCGTAGCCCAGGAATGACGTTGGATCGCATACCGCTATTCGTTTGATCTGTTGTAATCACTGCGGTTTTGTTATTGCTTTCTCTCCCGGTTCTCGCGGGGGCGATTACTATGCTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGACCCAGCGGGGGGAGGAGACCCAATTTTATACCA
#> 9                                               TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATCTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 10                                              TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCACTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 11                                              TGATCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGATGACCAAATTTATAACGTTATTGTTACTGCCCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 12                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 13                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGATGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGCATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 14                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCACCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGTGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTATGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCCCTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 15                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATACGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 16                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCACCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGGCCAGGAATAACATTAGACCGTATGCCACTATTTGTATGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCCCTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 17                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTCGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTATACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 18                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCATTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 19                                              TGGTCGGCACTTCACTCAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTTATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCTCCTTCGCTGACCTTGTTGGTGTCGAGCAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCCTTAGCTGCCAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGGAGTCCAGGAATAACGTTAGACCGCATACCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTCTCACTCCCTGTTCTAGCGGGGGCAATCACCATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGTGACCCAATCCTATACCA
#> 20                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATTACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTCCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 21                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGTTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 22                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCCCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 23                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCTATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 24                                              TGGTGGGGACATCGCTCAGTCTTTTAATTCGAGCCGAGCTTGGTAACCCTGGGTCCTTAATTGGTGACGACCAGATTTACAACGTTATTGTTACTGCTCACGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGGGGGTTTGGTAATTGGCTTGTTCCACTTATATTAGGGGCCCCCGACATAGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCTCCTTCCCTTACACTACTTGTGTCGAGCAGTGTGGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGGGCTTCAGTGGATTTCGCAATTTTTTCGTTACACCTGGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAACTTCATTACAACTGTAATTAATATGCGTAGCCCAGGAATGACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATCACTGCGGTTTTGTTATTGCTTTCTCTCCCGGTTCTCGCGGGGGCGATTACCATGCTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGACCCAGCGGGGGGAGGAGACCCAATTTTATACCA
#> 25                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 26                                              TGGTCGGCACTTCACTCAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTTATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCTCCTTCGCTGACCTTGTTGGTGTCGAGCAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCCAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGGAGTCCAGGAATAACGTTAGACCGCATACCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTCTCACTCCCTGTTCTAGCGGGGGCAATCACCATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGTGACCCAATCCTATACCA
#> 27                                              TGGTGGGGACATCGCTCAGTCTTTTAATTCGAGCCGAGCTTGGTAACCCTGGGTCCTTAATTGGTGACGACCAGATTTACAACGTTATTGTTACTGCTCACGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGGGGGTTTGGTAATTGGCTTGTTCCACTTATATTAGGGGCCCCCGACATAGCATTTCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCTCCTTCCCTTACACTACTTGTGTCGAGCAGTGTGGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGGGCTTCAGTGGATTTCGCAATTTTTTCGTTACACCTGGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAACTTCATTACAACTGTAATTAATATGCGTAGCCCAGGAATGACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATCACTGCGGTTTTGTTATTGCTTTCTCTCCCGGTTCTCGCGGGGGCGATTACCATGCTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGACCCAGCGGGGGGAGGAGACCCAATTTTATACCA
#> 28                                              TGGTGGGGACTTCACTGAGCCTGCTAATTCGAGCCGAACTTGGAAACCCGGGTTCTTTAATTGGGGATGACCAGATTTACAACGTGATTGTTACTGCCCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAATTGGTTAGTGCCTTTAATACTAGGGGCCCCAGACATGGCCTTTCCTCGTATAAACAATATAAGTTTTTGGTTGTTGCCGCCTTCGTTAACCTTGCTTGTGTCAAGTAGAATTGTAGATGTTGGTGCGGGAACGGGATGGACGGTGTACCCACCGCTTTCAGCCAATATCGCTCATGGTGGGGCATCTGTCGATTTTGCTATTTTTTCTCTACATTTGGCAGGGGTGTCTTCTATCTTAGGGGCTGTAAATTTTATTACAACTGTTGTTAACATGCGTAGCCCTGGAATGACTCTGGACCGAATACCCCTGTTTGTATGGTCTGTAGTAATTACCGCTGTTCTCTTACTCTTGTCGCTGCCGGTGCTGGCAGGAGCTATTACTATGTTACTAACGGATCGTAATTTGAATACCTCGTTTTTTGACCCTGCAGGTGGGGGTGATCCTATTTTGTACCA
#> 29 TACCCTTTATTTTATTTTTGGTGCGTGGTCGGGTATGGTGGGGACTTCACTGAGCCTGCTAATTCGAGCCGAACTTGGAAACCCGGGTTCTTTAATTGGGGATGACCAGATTTACAACGTGATTGTTACTGCCCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAATTGGTTAGTGCCTTTAATACTAGGGGCCCCAGACATGGCCTTTCCTCGTATAAACAATATAAGTTTTTGGTTGTTGCCGCCTTCGTTAACCTTGCTTGTGTCAAGTAGAATTGTAGATGTTGGTGCGGGAACGGGATGGACGGTGTACCCACCGCTTTCAGCCAATATCGCTCATGGTGGGGCATCTGTCGATTTTGCTATTTTTTCTCTACATTTGGCAGGGGTGTCTTCTATCTTAGGGGCTGTAAATTTTATTACAACTGTTGTTAACATGCGTAGCCCTGGAATGACTCTGGACCGAATACCCCTGTTTGTATGGTCTGTAGTAATTACCGCTGTTCTCTTACTCCTGTCGCTGCCGGTGCTGGCAGGAGCTATTACTATGTTACTAACGGATCGTAATTTGAATACCTCGTTTTTTGACCCTGCAGGTGGGGGTGATCCTATTTTGTACCAACATTTATTT
#> 30                                              TGGTGGGGACTTCACTGAGCCTGCTAATTCGAGCCGAACTTGGAAACCCGGGTTCTTTAATTGGGGATGACCAGATTTACAACGTGATTGTTACTGCCCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGAGGGTTTGGCAATTGGTTAGTGCCTTTAATACTAGGGGCCCCAGACATGGCCTTTCCTCGTATAAACAATATAAGTTTTTGGTTGTTGCCGCCTTCGTTAACCTTGCTTGTGTCAAGTAGAATTGTAGATGTTGGTGCGGGAACGGGATGGACGGTGTACCCACCGCTTTCAGCCAATATCGCTCATGGTGGGGCATCTGTCGATTTTGCTATTTTTTCTCTACATTTGGCAGGGGTGTCTTCTATCTTAGGGGCTGTAAATTTTATTACAACTGTTGTTAACATGCGTAGCCCTGGAATGACTCTGGACCGAATACCCCTGTTTGTATGGTCTGTAGTAATTACCGCTGTTCTCTTACTCCTGTCGCTGCCGGTGTTGGCAGGAGCTATTACTATGTTACTAACGGATCGTAATTTGAATACCTCGTTTTTTGACCCTGCAGGTGGGGGTGATCCTATTTTGTACCA
#> 31                                                TACTTTATACTTCATTTTTGGTGCATGAGCTGGAATAGTGGGCACTTCTTTGAGTTTATTAATTCGTGCAGAGCTAGGTAATCCTGGTTCTTTAATTGGTGATGACCAGATTTATAATGTTATTGTTACTGCCCATGCTTTTATTATAATTTTCTTTATAGTTATACCAATTATGATTGGTGGTTTTGGTAATTGATTAGTTCCTCTTATATTAGGAGCTCCTGATATAGCTTTTCCTCGTATAAATAATATAAGTTTTTGATTATTGCCTCCTTCATTGACTTTATTAGTATCAAGTAGTTTAGTAGATATAGGAGCTGGTACGGGTTGAACTGTATATCCTCCTTTAGCGGCTAATATTGCTCATGGAGGGTCATCAGTTGATTACGCTATTTTTTCTTTACATTTAGCTGGGGTATCTTCTATTTTAGGTGCTGTAAATTTTATTACAACAGTAATTAACATGCGTAGCCCTGGTATAACTTTAGATCGTATTCCTTTATTTGTATGATCTGTTGTAATTACTGCTGTGTTATTGCTTTTATCGCTACCTGTATTAGCCGGTGCTATTACTATACTTTTAACCGATCGTAATCTTAATACTTCTTTTT
#> 32    TACTTTATACTTCATTTTTGGTGCATGAGCTGGTATAGTGGGCACTTCATTGAGTTTATTAATTCGTGCGGAACTAGGGAATCCTGGATCTTTAATTGGTGATGATCAAATTTATAATGTTATTGTTACTGCTCACGCTTTTATTATAATTTTCTTTATGGTTATACCAATCATGATTGGAGGTTTTGGTAATTGATTAGTACCTCTTATATTAGGAGCACCTGATATGGCCTTCCCTCGTATAAATAATATAAGTTTTTGATTATTACCTCCTTCATTAACTTTATTAGTATCAAGTAGTTTAGTTGATATGGGAGCCGGTACAGGCTGAACAGTTTATCCTCCTTTAGCTGCAAATATTGCTCATGGAGGGTCATCCGTTGATTATGCTATTTTTTCTTTACATTTAGCTGGGGTATCTTCTATTTTAGGTGCTGTAAATTTTATTACAACAGTAATTAATATGCGTAGTCCGGGTATAACTTTAGATCGAATTCCTTTATTTGTATGATCTGTTGTAATTACTGCTGTTTTATTATTATTATCACTACCTGTGTTAGCTGGTGCTATTACTATACTTTTAACTGATCGTAATCTTAATACTTCTTTTTTTGATCCTGCTGGAGGGGGTGATCCTATTTTATACCAACATTTA
#> 33                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTTTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 34                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCACCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTATGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCCCTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 35                                              TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTTCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCGCTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 36                                              TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATCATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTATATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCGCTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 37                                              TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCGCTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 38                                              TGGTGGGGACATCGCTCAGTCTTTTAATTCGAGCCGAGCTTGGTAACCCTGGGTCCTTAATTGGTGACGACCAGATTTACAACGTTATTGTTACTGCTCACGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGGGGGTTTGGTAATTGGCTTGTTCCCCTTATATTAGGGGCCCCCGACATAGCATTTCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCTCCTTCCCTTACACTACTTGTGTCGAGCAGTGTGGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCGCATGGGGGGGCTTCAGTGGATTTCGCAATTTTTTCGTTACACCTGGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAACTTCATTACAACTGTAATTAATATGCGTAGCCCAGGAATGACGTTGGATCGCATACCGCTATTCGTTTGATCTGTTGTAATCACTGCGGTTTTGTTATTGCTTTCTCTCCCGGTTCTCGCGGGGGCGATTACTATGCTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGACCCAGCGGGGGGAGGAGACCCAATTTTATACCA
#> 39                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATCTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 40                                              TGGTGGGGACCTCCCTCAGTCTTTTAATTCGAGCCGAGCTTGGCAATCCTGGGTCTCTAATTGGTGACGACCAAATTTACAACGTTATTGTCACTGCTCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGCGGGTTTGGTAATTGGCTTGTTCCGCTCATGTTAGGAGCCCCAGATATGGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCCCCTTCGCTCACACTGCTTGTATCAAGCAGTGTTGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGTGCTTCTGTGGATTTCGCAATTTTTTCACTACACCTGGCGGGGGTGTCTTCAATTTTGGGCGCTGTAAACTTCATTACAACTGTAATTAATATGCGCAGCCCAGGAATAACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATTACTGCGGTTTTATTACTACTTTCTCTACCGGTTCTTGCCGGGGCAATTACTATACTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGATCCGGCGGGGGGAGGAGACCCAATTTTATACCA
#> 41                                              TGATCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGATGACCAAATTTATAACGTTATTGTTACTGCCCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 42                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 43                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGATGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGCATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 44                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCACCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGTGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTATGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCCCTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 45                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATACGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 46                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCACCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGGCCAGGAATAACATTAGACCGTATGCCACTATTTGTATGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCCCTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 47                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTCGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTATACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 48                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCATTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 49                                              TGGTCGGCACTTCACTCAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTTATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCTCCTTCGCTGACCTTGTTGGTGTCGAGCAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCCTTAGCTGCCAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGGAGTCCAGGAATAACGTTAGACCGCATACCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTCTCACTCCCTGTTCTAGCGGGGGCAATCACCATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGTGACCCAATCCTATACCA
#> 50                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATTACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTCCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 51                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGTTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 52                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCCCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 53                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCTATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 54                                              TGGTGGGGACATCGCTCAGTCTTTTAATTCGAGCCGAGCTTGGTAACCCTGGGTCCTTAATTGGTGACGACCAGATTTACAACGTTATTGTTACTGCTCACGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGGGGGTTTGGTAATTGGCTTGTTCCACTTATATTAGGGGCCCCCGACATAGCATTCCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCTCCTTCCCTTACACTACTTGTGTCGAGCAGTGTGGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGGGCTTCAGTGGATTTCGCAATTTTTTCGTTACACCTGGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAACTTCATTACAACTGTAATTAATATGCGTAGCCCAGGAATGACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATCACTGCGGTTTTGTTATTGCTTTCTCTCCCGGTTCTCGCGGGGGCGATTACCATGCTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGACCCAGCGGGGGGAGGAGACCCAATTTTATACCA
#> 55                                              TGGTCGGCACTTCACTTAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAGATTTATAATGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTCATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCCCCTTCGCTGACCTTGTTGGTGTCGAGTAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACGGTGTACCCTCCTTTAGCTGCGAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGTAGTCCAGGAATAACATTAGACCGTATGCCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTATCACTCCCTGTTCTAGCGGGGGCAATCACTATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGCGACCCAATCTTATACCA
#> 56                                              TGGTCGGCACTTCACTCAGTCTTTTAATTCGAGCCGAGCTTGGGAACCCTGGTTCATTAATTGGCGACGACCAAATTTATAACGTTATTGTTACTGCTCATGCTTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGTTTTGGCAACTGGCTGGTGCCGCTTATGTTGGGTGCGCCTGATATGGCGTTTCCACGTATAAATAATATGAGTTTTTGGTTGTTACCTCCTTCGCTGACCTTGTTGGTGTCGAGCAGAGTAGTAGATGTGGGGGCTGGCACGGGGTGAACAGTGTACCCTCCTTTAGCTGCCAACATCGCGCATGGGGGCGCCTCAGTGGATTTTGCGATTTTTTCATTACATTTAGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAATTTTATCACTACTGTAGTTAACATGCGGAGTCCAGGAATAACGTTAGACCGCATACCACTATTTGTCTGGTCTGTTGTAATCACTGCGGTTTTATTATTGCTCTCACTCCCTGTTCTAGCGGGGGCAATCACCATACTTTTAACGGATCGTAATTTAAACACTTCGTTTTTTGACCCGGCGGGGGGTGGTGACCCAATCCTATACCA
#> 57                                              TGGTGGGGACATCGCTCAGTCTTTTAATTCGAGCCGAGCTTGGTAACCCTGGGTCCTTAATTGGTGACGACCAGATTTACAACGTTATTGTTACTGCTCACGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATCGGGGGGTTTGGTAATTGGCTTGTTCCACTTATATTAGGGGCCCCCGACATAGCATTTCCCCGCATAAATAATATAAGTTTCTGGTTGTTACCTCCTTCCCTTACACTACTTGTGTCGAGCAGTGTGGTTGATGTAGGGGCGGGCACTGGGTGGACCGTGTATCCCCCCTTGGCTGCCAACATCGCACATGGGGGGGCTTCAGTGGATTTCGCAATTTTTTCGTTACACCTGGCGGGGGTGTCTTCAATTTTGGGTGCTGTAAACTTCATTACAACTGTAATTAATATGCGTAGCCCAGGAATGACGTTGGATCGCATACCGCTATTCGTTTGGTCTGTTGTAATCACTGCGGTTTTGTTATTGCTTTCTCTCCCGGTTCTCGCGGGGGCGATTACCATGCTGCTAACTGACCGTAATTTAAATACTTCATTTTTTGACCCAGCGGGGGGAGGAGACCCAATTTTATACCA
#> 58                                              TGGTGGGGACTTCACTGAGCCTGCTAATTCGAGCCGAACTTGGAAACCCGGGTTCTTTAATTGGGGATGACCAGATTTACAACGTGATTGTTACTGCCCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAATTGGTTAGTGCCTTTAATACTAGGGGCCCCAGACATGGCCTTTCCTCGTATAAACAATATAAGTTTTTGGTTGTTGCCGCCTTCGTTAACCTTGCTTGTGTCAAGTAGAATTGTAGATGTTGGTGCGGGAACGGGATGGACGGTGTACCCACCGCTTTCAGCCAATATCGCTCATGGTGGGGCATCTGTCGATTTTGCTATTTTTTCTCTACATTTGGCAGGGGTGTCTTCTATCTTAGGGGCTGTAAATTTTATTACAACTGTTGTTAACATGCGTAGCCCTGGAATGACTCTGGACCGAATACCCCTGTTTGTATGGTCTGTAGTAATTACCGCTGTTCTCTTACTCTTGTCGCTGCCGGTGCTGGCAGGAGCTATTACTATGTTACTAACGGATCGTAATTTGAATACCTCGTTTTTTGACCCTGCAGGTGGGGGTGATCCTATTTTGTACCA
#> 59 TACCCTTTATTTTATTTTTGGTGCGTGGTCGGGTATGGTGGGGACTTCACTGAGCCTGCTAATTCGAGCCGAACTTGGAAACCCGGGTTCTTTAATTGGGGATGACCAGATTTACAACGTGATTGTTACTGCCCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGTGGGTTTGGCAATTGGTTAGTGCCTTTAATACTAGGGGCCCCAGACATGGCCTTTCCTCGTATAAACAATATAAGTTTTTGGTTGTTGCCGCCTTCGTTAACCTTGCTTGTGTCAAGTAGAATTGTAGATGTTGGTGCGGGAACGGGATGGACGGTGTACCCACCGCTTTCAGCCAATATCGCTCATGGTGGGGCATCTGTCGATTTTGCTATTTTTTCTCTACATTTGGCAGGGGTGTCTTCTATCTTAGGGGCTGTAAATTTTATTACAACTGTTGTTAACATGCGTAGCCCTGGAATGACTCTGGACCGAATACCCCTGTTTGTATGGTCTGTAGTAATTACCGCTGTTCTCTTACTCCTGTCGCTGCCGGTGCTGGCAGGAGCTATTACTATGTTACTAACGGATCGTAATTTGAATACCTCGTTTTTTGACCCTGCAGGTGGGGGTGATCCTATTTTGTACCAACATTTATTT
#> 60                                              TGGTGGGGACTTCACTGAGCCTGCTAATTCGAGCCGAACTTGGAAACCCGGGTTCTTTAATTGGGGATGACCAGATTTACAACGTGATTGTTACTGCCCATGCGTTTATTATAATTTTTTTTATAGTGATACCAATTATAATTGGAGGGTTTGGCAATTGGTTAGTGCCTTTAATACTAGGGGCCCCAGACATGGCCTTTCCTCGTATAAACAATATAAGTTTTTGGTTGTTGCCGCCTTCGTTAACCTTGCTTGTGTCAAGTAGAATTGTAGATGTTGGTGCGGGAACGGGATGGACGGTGTACCCACCGCTTTCAGCCAATATCGCTCATGGTGGGGCATCTGTCGATTTTGCTATTTTTTCTCTACATTTGGCAGGGGTGTCTTCTATCTTAGGGGCTGTAAATTTTATTACAACTGTTGTTAACATGCGTAGCCCTGGAATGACTCTGGACCGAATACCCCTGTTTGTATGGTCTGTAGTAATTACCGCTGTTCTCTTACTCCTGTCGCTGCCGGTGTTGGCAGGAGCTATTACTATGTTACTAACGGATCGTAATTTGAATACCTCGTTTTTTGACCCTGCAGGTGGGGGTGATCCTATTTTGTACCA
#>    markercode
#> 1      COI-5P
#> 2      COI-5P
#> 3      COI-5P
#> 4      COI-5P
#> 5      COI-5P
#> 6      COI-5P
#> 7      COI-5P
#> 8      COI-5P
#> 9      COI-5P
#> 10     COI-5P
#> 11     COI-5P
#> 12     COI-5P
#> 13     COI-5P
#> 14     COI-5P
#> 15     COI-5P
#> 16     COI-5P
#> 17     COI-5P
#> 18     COI-5P
#> 19     COI-5P
#> 20     COI-5P
#> 21     COI-5P
#> 22     COI-5P
#> 23     COI-5P
#> 24     COI-5P
#> 25     COI-5P
#> 26     COI-5P
#> 27     COI-5P
#> 28     COI-5P
#> 29     COI-5P
#> 30     COI-5P
#> 31     COI-5P
#> 32     COI-5P
#> 33     COI-5P
#> 34     COI-5P
#> 35     COI-5P
#> 36     COI-5P
#> 37     COI-5P
#> 38     COI-5P
#> 39     COI-5P
#> 40     COI-5P
#> 41     COI-5P
#> 42     COI-5P
#> 43     COI-5P
#> 44     COI-5P
#> 45     COI-5P
#> 46     COI-5P
#> 47     COI-5P
#> 48     COI-5P
#> 49     COI-5P
#> 50     COI-5P
#> 51     COI-5P
#> 52     COI-5P
#> 53     COI-5P
#> 54     COI-5P
#> 55     COI-5P
#> 56     COI-5P
#> 57     COI-5P
#> 58     COI-5P
#> 59     COI-5P
#> 60     COI-5P