Plot an interactive map

This functions generate an interactive maps showing the location of the records of a reference database. Note that only records with latitude and longitude data will be displayed.

refdb_plot_map(x)

Arguments

x

a reference database.

Value

An interactive map object from the leaflet package.

Examples

lib <- read.csv(system.file("extdata", "baetidae_bold.csv", package = "refdb"))
lib <- refdb_set_fields_BOLD(lib)
lib <- refdb_set_fields(lib, latitude = "lat", longitude = "lon")
refdb_plot_map(lib)
#> Warning: Data contains 28 rows with either missing or invalid lat/lon values and will be ignored